+++ ChangeLog.md
@@ -1,3 +1,13 @@
 # Changelog for vehicle
 
-## Unreleased changes
+## Version 0.0.2.0
+
+- Added `dataset` declarations.
+- Added `type` declarations.
+- Renamed kind of type-level naturals to `Dim`.
+- Renamed type of naturals to `Int`.
+- Changed syntax of cons to infix `::`.
+
+## Version 0.0.1.0
+
+- First version presented internally on 2021/03/08.